% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlemap_json_to_string.R
\name{googlemap_json_to_string}
\alias{googlemap_json_to_string}
\title{Convert Googlemap JSON style sheet to ggmap API format}
\usage{
googlemap_json_to_string(style_list)
}
\arguments{
\item{style_list}{JSON format object, easiest to build one of these using snazzymaps.com or similar service}
}
\value{
Base R heatmap plot with row labels on left instead of right side
}
\description{
Takes a JSON style guide for Google Maps and returns a Google Map API string for styling
}
\examples{
mapstyle <- rjson::fromJSON(file = "./R/mapstyle3.json") \%>\%
            googlemap_json_to_string(.)
ggmap::get_googlemap(style=mapstyle)

}
