% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_left.R
\name{heatmap_left}
\alias{heatmap_left}
\title{Basic heatmap with row labels on left}
\usage{
heatmap_left(
  x,
  Rowv = NULL,
  Colv = if (symm) "Rowv" else NULL,
  distfun = dist,
  hclustfun = hclust,
  reorderfun = function(d, w) reorder(d, w),
  add.expr,
  symm = FALSE,
  revC = identical(Colv, "Rowv"),
  scale = c("row", "column", "none"),
  na.rm = TRUE,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  cexRow = 0.2 + 1/log10(nr),
  cexCol = 0.2 + 1/log10(nc),
  labRow = NULL,
  labCol = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  keep.dendro = FALSE,
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{x}{a numeric matrix}
}
\value{
Base R heatmap plot with row labels on left instead of right side
}
\description{
Returns a heatmap plot (from base R) with row labels on left instead of right side
}
\examples{
heatmap_left(matrix(rnorm(100),nrow=10))

}
