% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_bbdml.R
\name{multi_bbdml}
\alias{multi_bbdml}
\title{Basic heatmap with row labels on left}
\usage{
multi_bbdml(
  da_analysis,
  ps_object,
  mu_predictor,
  phi_predictor,
  seed = 123,
  taxlevels = 1:7
)
}
\arguments{
\item{da_analysis}{output from corncob::differentialTest()}

\item{ps_object}{phyloseq object}

\item{mu_predictor}{Variable in sample_data(ps_object) as character vector of length 1}

\item{phi_predictor}{Variable in sample_data(ps_object) as character vector of length 1}

\item{seed}{Random seed (integer) ... Default=123}

\item{taxlevels}{Numeric vector. Taxonomic levels to include in names. Default = 1:7. 1=Kingdom, 7=Species}
}
\value{
List object of bbdml models for all significant taxa found with corncob::differentialTest()
}
\description{
Automate some aspects of bbdml for taxa found to be differentially abundant via differentialTest(). multi_bbdml() takes the output from corncob::differentialTest() along with a single mu and phi predictor variable and returns a list of bbdml objects for each taxon that had significant differential abundance. This is different from differentialTest() in that not just the summary objects are returned. This allows for plots to be generated programmatically using the plot_multi_bbdml() function, which takes the output from multi_bbdml() and returns a single bbdml plot for each significant taxon.
}
\examples{
bbs <- multi_bbdml(da_analysis = da_analysis,
                   ps_object = ps_family,
                   mu_predictor = "FireTreatment",
                   phi_predictor = "FireTreatment",
                   taxlevels = c(1,2,3,4))

}
