% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_bbdml.R
\name{plot_multi_bbdml}
\alias{plot_multi_bbdml}
\title{Autmoates plotting multiple bbdml plots from taxa identified as significantly different by multi_bbdml()}
\usage{
plot_multi_bbdml(
  bbdml_list,
  color = "none",
  obj_basename = "bbdml_plot_",
  pointsize = 1,
  whichtaxa = 1:length(bbdml_list)
)
}
\arguments{
\item{bbdml_list}{output list object from multi_bbdml()}

\item{color}{Character vector of length 1. Name of a variable found in the bbdml_list list object to color points by. Default = 'none'.}

\item{obj_basename}{Character to begin name of returned ggplot objects. Default = "bbdml_plot_"}

\item{pointsize}{size of points in plots. Default = 1}

\item{whichtaxa}{Numeric vector indicating the elements of the bbdml_list to plot. Defaults to all.}
}
\value{
Base R heatmap plot with row labels on left instead of right side
}
\description{
Returns a series of ggplot objects to .GlobalEnv, named by the obj_basename parameter. This function is called for its side effects. Yeah, that's bad and maybe it should return them as a list. Maybe I'll change that later.
}
\examples{
plot_multi_bbdml(bbdml_list = bbs,pointsize=3,color = "FireTreatment",whichtaxa = 1:2)
library(patchwork)
bbdml_plot_1 / bbdml_plot_2

}
